Backend/requirements.txt
flask
google-genai
flask-cors
python-dotenv
